fixed many things

+++ server/controller/assignmentController.js
@@ -1,6 +1,9 @@
 // Import custom files
 const AssignmentClient = require("database").AssignmentClient;
 const AssignmentErrorHandler = require("../error/assignmentErrorHandler");
+const GeneralErrorHandler = require("../error/generalErrorHandler");
+
+const isValidUUID = require("../helper/helper").isValidUUID;
 
 const assignmentClient = new AssignmentClient();
 
@@ -14,6 +17,9 @@ const assignmentClient = new AssignmentClient();
 const createAssignment = async (req, res) => {
   try {
     const { name, points, num_of_attempts, deadline } = req.body;
+    if (!name || !points || !num_of_attempts || !deadline) throw new GeneralErrorHandler('GEN_102');
+    if (points>10 || points<1) throw new AssignmentErrorHandler('ASSGN_102')
+    if (deadline <= new Date().toISOString()) throw new AssignmentErrorHandler('ASSGN_102')
     const payload = {
       name: name,
       points: points,
@@ -22,47 +28,75 @@ const createAssignment = async (req, res) => {
     };
     const assignment = await assignmentClient.createAssignment(
       payload,
-      req.user.username,
+      req.user.id,
     );
     res.status(201).send(assignment);
   } catch (error) {
     console.error("Error creating assignment:", error);
-    res.status(400).send();
+    if (!error.statusCode) error.statusCode = 500;
+    res.status(error.statusCode).send();
   }
 };
 
 const deleteAssignment = async (req, res) => {
   try {
+    if (!isValidUUID(req.params.id)) throw new AssignmentErrorHandler('ASSGN_103');
+
     const isDeleted = await assignmentClient.deleteAssignment(req.params.id);
-    if (isDeleted) res.status(200).send();
-    else res.status(201).send();
+    if (!isDeleted) throw new AssignmentErrorHandler('ASSGN_101')
+    res.status(204).send();
   } catch (error) {
     console.error("Error deleting assignment:", error);
-    res.status(503).send();
+    if (!error.statusCode) error.statusCode = 500;
+    res.status(error.statusCode).send();
   }
 };
 
 const getAssignment = async (req, res) => {
   try {
+    if (!isValidUUID(req.params.id)) throw new AssignmentErrorHandler('ASSGN_103');
+    
     const assignment = await assignmentClient.getAssignment(req.params.id);
     if (!assignment) throw new AssignmentErrorHandler("ASSGN_101");
+
     res.status(200).send(assignment.toJSON());
   } catch (error) {
     console.error("Error getting assignment:", error);
-    res.status(404).send(error);
+    if (!error.statusCode) error.statusCode = 500;
+    res.status(error.statusCode).send();
   }
 };
 
 const getAllAssignment = async (req, res) => {
   try {
+    if (Object.keys(req.body).length > 0) throw new GeneralErrorHandler('GEN_102');
+
     const assignments = await assignmentClient.getAllAssignment();
     assignments.forEach((assignment) => {
       return assignment.toJSON();
     });
     res.status(200).send(assignments);
   } catch (error) {
-    console.error("Error fetching assignment:", error);
-    res.status(503).send();
+    console.error("Error getting assignment:", error);
+    if (!error.statusCode) error.statusCode = 500;
+    res.status(error.statusCode).send();
+  }
+};
+
+const updateAssignment = async (req, res) => {
+  try {
+    const assignmentId = req.params.id;
+    if (!assignmentId) throw new AssignmentErrorHandler('ASSGN_103');
+
+    const assignment = Object.keys(req.body).reduce((result, key) => (req.body[key] !== null ? { ...result, [key]: req.body[key] } : result), {});
+    if (!assignment) throw new GeneralErrorHandler('GEN_102');
+
+    const updatedAssignment = await assignmentClient.updateAssignment(assignment, assignmentId, req.user.id);
+    res.status(200).send(updatedAssignment);
+  } catch (error) {
+    console.error("Error updating assignment:", error);
+    if (!error.statusCode) error.statusCode = 500;
+    res.status(error.statusCode).send();
   }
 };
 
@@ -71,4 +105,5 @@ module.exports = {
   deleteAssignment,
   getAssignment,
   getAllAssignment,
+  updateAssignment,
 };
